Fix JWT problem

+++ course-02/exercises/udacity-c2-restapi/src/controllers/v0/users/routes/auth.router.ts
@@ -22,7 +22,7 @@ async function comparePasswords(plainTextPassword: string, hash: string): Promis
 }
 
 function generateJWT(user: User): string {
-    return jwt.sign(user, config.jwt.secret)
+    return jwt.sign(user.short(), config.jwt.secret)
 }
 
 export function requireAuth(req: Request, res: Response, next: NextFunction) {
@@ -35,6 +35,7 @@ export function requireAuth(req: Request, res: Response, next: NextFunction) {
         return res.status(401).send({ message: 'Malformed token.' });
     }
     
+    // token of form 'bearer <token>'
     const token = token_bearer[1];
 
     // doesn't expire?
